third_party/WebKit/LayoutTests/http/tests/htmlimports/import-async.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link id="target" async rel="import" href="../../../resources/redirect.php?delay=100&url=../htmlimports/resources/hello.html">
</head>
<body>
<script>
(function() {
var test = async_test('With the async attribute, link element don\'t block the script execution.');

test.step(function() {
    assert_true(window.target.import === null, "Script goes without waiting the link loaded");
});

window.target.onload = function() {
    test.step(function() {
        assert_true(window.target.import instanceof HTMLDocument, "Ensureing that the import being loaded eventually.");
        test.done();
    });
};
})();
</script>
</body>
</html>
